set up rollup, generate
dist\core.iife.js
dist\core.umd.js
dist\core.js
drop grunt uglify, amd, concat, release
