+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "core.js",
-  "version": "0.7.3",
+  "version": "0.8.0",
   "description": "Lightweight framework for scalable applications",
   "author": "Mauricio Soares",
   "homepage": "https://github.com/msodeveloper/core.js",
@@ -11,7 +11,7 @@
   },
   "main": "dist/core.js",
   "scripts": {
-    "test": "grunt ci"
+      "bundle": "rollup --config tools/rollup.config.js"
   },
   "bugs": {
     "url": "https://github.com/msodeveloper/core.js/issues"
@@ -23,15 +23,6 @@
     "scalable"
   ],
   "devDependencies": {
-    "grunt": "^0.4.5",
-    "grunt-bump": "^0.3.0",
-    "grunt-contrib-concat": "^0.5.0",
-    "grunt-contrib-jasmine": "^0.8.0",
-    "grunt-contrib-jshint": "^0.10.0",
-    "grunt-contrib-uglify": "^0.6.0",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-coveralls": "^1.0.0",
-    "grunt-template-jasmine-istanbul": "^0.3.0",
-    "grunt-umd": "^2.3.3"
+    "rollup": "^0.66.6"
   }
 }
