+++ tools/gruntdep.json
@@ -1,0 +1,17 @@
+{
+    "scripts": {
+        "test": "grunt ci"
+    },
+    "devDependencies": {
+      "grunt": "^0.4.5",
+      "grunt-bump": "^0.3.0",
+      "grunt-contrib-concat": "^0.5.0",
+      "grunt-contrib-jasmine": "^0.8.0",
+      "grunt-contrib-jshint": "^0.10.0",
+      "grunt-contrib-uglify": "^0.6.0",
+      "grunt-contrib-watch": "^0.6.1",
+      "grunt-coveralls": "^1.0.0",
+      "grunt-template-jasmine-istanbul": "^0.3.0",
+      "grunt-umd": "^2.3.3"
+    }
+}
